# backend/openapi.json
{
  "components": {
    "schemas": {
      "AuthChallengeRead": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          }
        },
        "required": [
          "message",
          "nonce",
          "expires_at"
        ],
        "title": "AuthChallengeRead",
        "type": "object"
      },
      "AuthChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "chain_id": {
            "title": "Chain Id",
            "type": "integer"
          },
          "wallet_address": {
            "maxLength": 42,
            "minLength": 42,
            "title": "Wallet Address",
            "type": "string"
          }
        },
        "required": [
          "wallet_address",
          "chain_id"
        ],
        "title": "AuthChallengeRequest",
        "type": "object"
      },
...
      "AuthVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "maxLength": 4096,
            "minLength": 80,
            "title": "Message",
            "type": "string"
          },
          "signature": {
            "pattern": "^0x[0-9a-fA-F]{130}$",
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "message",
          "signature"
        ],
        "title": "AuthVerifyRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Off-chain metadata, wallet authentication, and Base Sepolia event projection.",
    "title": "TrustWork API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/challenge": {
      "post": {
        "operationId": "auth_challenge_auth_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthChallengeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auth Challenge"
      }
    },
    "/auth/verify": {
      "post": {
        "operationId": "auth_verify_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auth Verify"
      }
    },
...
    "/disputes/{dispute_id}/evidence": {
      "get": {
        "operationId": "list_dispute_evidence_disputes__dispute_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceRead"
                  },
                  "title": "Response List Dispute Evidence Disputes  Dispute Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dispute Evidence"
      }
    },
...
    "/evidence": {
      "post": {
        "operationId": "create_evidence_evidence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Evidence"
      }
    },
...
    "/indexer/reconcile": {
      "post": {
        "operationId": "reconcile_indexer_indexer_reconcile_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Indexer-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Indexer-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexerPollRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Indexer"
      }
    },
    "/indexer/sync": {
      "post": {
        "operationId": "sync_indexer_indexer_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexerSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexerPollRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sync Indexer"
      }
    },
...
    "/jobs/prepare": {
      "post": {
        "operationId": "prepare_job_jobs_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPrepareRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Prepare Job"
      }
    },
    "/jobs/{job_id}/evidence": {
      "get": {
        "operationId": "list_job_evidence_jobs__job_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceRead"
                  },
                  "title": "Response List Job Evidence Jobs  Job Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Job Evidence"
      }
    },
    "/jobs/{onchain_job_id}": {
      "get": {
        "operationId": "get_job_jobs__onchain_job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "onchain_job_id",
            "required": true,
            "schema": {
              "title": "Onchain Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/matches/{wallet_address}": {
      "get": {
        "operationId": "list_matches_matches__wallet_address__get",
        "parameters": [
          {
            "in": "path",
            "name": "wallet_address",
            "required": true,
            "schema": {
              "title": "Wallet Address",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchRead"
                  },
                  "title": "Response List Matches Matches  Wallet Address  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Matches"
      }
    },
...
    "/reputation/refresh-all": {
      "post": {
        "operationId": "refresh_all_reputations_reputation_refresh_all_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Indexer-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Indexer-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReputationRead"
                  },
                  "title": "Response Refresh All Reputations Reputation Refresh All Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh All Reputations"
      }
    },
    "/reputation/{wallet_address}": {
      "get": {
        "operationId": "get_reputation_reputation__wallet_address__get",
        "parameters": [
          {
            "in": "path",
            "name": "wallet_address",
            "required": true,
            "schema": {
              "title": "Wallet Address",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reputation"
      }
    },
    "/reputation/{wallet_address}/refresh": {
      "post": {
        "operationId": "refresh_reputation_reputation__wallet_address__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "wallet_address",
            "required": true,
            "schema": {
              "title": "Wallet Address",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refresh Reputation"
      }
    },
    "/swipes": {
      "post": {
        "operationId": "create_swipe_swipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwipeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwipeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Swipe"
      }
    },
...
  }
}